[fix]: [CDS-40179]: Added timescaleModule (#38769)
* added timescaleModule

* removed unused imports

* removed else block
